+++ HISTORY.rst
@@ -3,12 +3,13 @@
 History
 -------
 
-1.0.0 (2018-10-10)
+1.0.0 (2018-10-12)
 ---------------------
 * Mutations can be chosen from a subset (pool) of all possible positions.
 * Replicates can be partitioned into multiple groups with each group sharing a pool of eligible positions.
 * Mutations can be either monomorphic or polymorphic.
 * Add the capability to generate VCF output files.
+* Add the capability to generate a metrics output file.
 * The installer creates an executable script called ``snpmutator``.
 * Insertions are now placed after the original reference position, not before, as in prior versions of
   snpmutator.  This means it will not be possible to exactly reproduce the results of prior versions